fix(controller): sleep until publisher's refresh interval
Deis' publisher component has a maximum latency of 10 seconds before
it refreshes its config from etcd. For a true zero-downtime
deployment, we want to make sure that the new containers are
templated in the router before performing the health check and/or
recycling the old containers.

The FIXME here would be to utilize confd --watch in the router and
in publisher, but that is blocked due to an issue upstream with
etcd[1].

[1]: https://github.com/coreos/etcd/issues/2679